+++ src/main.hpp
@@ -9,7 +9,7 @@
 
 #include <list>
 #include <pybind11/pybind11.h>
-#include <pybind11/stl_bind.h>
+// #include <pybind11/stl_bind.h>
 #include <unicode/errorcode.h>
 #include <unicode/locid.h>
 #include <unicode/unistr.h>